消息页面分页查询完成。多酒吧登录后，用户页面屏蔽删除按钮，显示用户为在酒吧法过消息的用户

+++ routes/message.js
@@ -7,26 +7,78 @@ const Token = require('../lib/Token')
 const auth = require('../middlewares/auth')
 const co = require('co')
 const router = express.Router()
-const {Message} = models
+const {Message, sequelize} = models
 
 router.get('/', auth, (req, res, next) => {
   const {id} = res.locals.barInfo
 
+  let limit = req.query['limit'] || 50
+  let offset = req.query['offset'] || 0
+
+  if (!/\d+/.test(limit) || !/\d+/.test(offset)) {
+    res.json({iRet: -1, msg: "参数错误"})
+    return
+  }
+
   co(function*() {
-    var messages = yield Message.findAll({
-      where: {
-        BarId: id
-      }
-    })
+    let count = yield Message.count({where: {BarId: id}})
 
-    messages = messages.map(o => o.get({plain: true}))
+    let _sql = "SELECT * FROM Messages WHERE (deletedAt IS NULL AND BarId = 1) ORDER BY createdAt DESC LIMIT 50"
 
-    res.render('message', {messages})
+    let messages = yield sequelize.query(_sql)
+
+    if (messages && messages[0]) {
+      res.render('message', {messages: messages[0], count: count})
+    } else
+      res.json({iRet: -1, msg: err})
+
+  }).catch((err) => {
+
+    res.json({iRet: -1, msg: err})
+
+  })
+})
+
+router.get('/query', auth, (req, res, next) => {
+  const {id} = res.locals.barInfo
+
+  let limit = req.query['limit'] || 50
+  let offset = req.query['offset'] || 0
+
+  if (!/\d+/.test(limit) || !/\d+/.test(offset)) {
+    res.json({iRet: -1, msg: "参数错误"})
+    return
+  }
+
+  co(function*() {
+    var count = yield Message.count({where: {BarId: id}})
+
+    let _sql = `SELECT * FROM Messages WHERE (deletedAt IS NULL AND BarId = 1) ORDER BY createdAt DESC LIMIT ${offset}, ${limit}`
+
+    let messages = yield sequelize.query(_sql)
+
+    if (messages && messages[0]) {
+
+      res.json({messages: messages[0], count: count})
+
+    } else {
+
+      res.json({iRet: -1, msg: err})
+    }
+  }).catch((err) => {
+
+    res.json({iRet: -1, msg: err})
+
   })
 })
 
 router.post('/del', (req, res, next) => {
   var id = req.body.id
+
+  if (!/\d+/.test(id)) {
+    res.json({iRet: -1, msg: "参数错误"})
+    return
+  }
 
   Message.destroy({
     where: {
